+++ src/main/java/com/esprit/tn/forum/service/IrecruiterFeedbackService.java
@@ -10,7 +10,7 @@ public interface IrecruiterFeedbackService {
 
     List<Feedback> retrieveAllFeedback();
 
-    Feedback addFeedback(Feedback f, Long idcondidacy);
+    Feedback addFeedback(Feedback f, int idcondidacy);
 
     Feedback updateFeedbackrecruiter(Feedback f);
